Refactor user merge classes across files
