+++ coldfront/core/user/utils_/merge_users/__init__.py
@@ -1,0 +1,6 @@
+from coldfront.core.user.utils_.merge_users.runner import UserMergeRunner
+
+
+__all__ = [
+    'UserMergeRunner',
+]
